vendor service api: reverse vendor field vs site field
In the previous revision, the format was site_field:vendor_field.
This however removes the ability for an entitlement to have multiple
vendor_fields mapped to the same site_field, which in some cases,
may be desired.

This commit breaks entitlements on the frontend as this is a breaking
API change.